backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
aiomysql==0.2.0
pydantic==2.9.2
python-dotenv==1.0.1
